HPCRUSHIN'IT-60 Adds random colors, population methods
rand(n) is memoized so that it never returns the same value for the
same 'n' twice in a row. The template assigns a color class to the
frame border that is randomly assigned when rendered.

Adds populateGrid() function that prints only large pictures as long
as there are more than 4 times as many important pictures as filler
pictures; then prints a 1-to-4 mix of important and filler pictures,
then prints only filler when the important pics run out.

Updates underscore to use 1.6.0 (which has the partition method)