Huge update: Adds bootstrap loader
All that's needed to load the photowall (having already loaded jQuery)
is to add a single div/section/whatever with the class
"photowall-container" and a data-layout tag of the layout json file,
followed by a script tag that sources "bootstrap.js".

Should be able to work with multiple photowall containers.